I don't even want to know....
The previous way of checking the child status was randomly failing.

I can't really explain why, but it wasn't rasing, which is supposed
to mean the call was valid, but the `$?` aka $CHILD_STATUS special
variable wasn't properly set sometimes.

It seems to work that way even if I have no explainations of why.