Reorder active_tasks index
That index was added for the following query:

```sql
SELECT  `tasks`.*
FROM `tasks`
WHERE `tasks`.`stack_id` = ?
AND `tasks`.`status` IN ("pending", "running", "aborting")
AND `tasks`.`allow_concurrency` = false
ORDER BY `tasks`.`id` DESC
LIMIT 1
```

Since the criteria on `status` is a range, for MySQL to use that index
`status` need to be the last element of the index.

I don't know about how Postgres was performing on that query,
but I assume that limitation applies to it as well.

Thanks to @sroysen for the fix.