Add information about who aborted a deploy
We had a problem because someone aborted a deploy and triggered a
rollback and we had no way to know who triggered the abort.

Looking at the code I found that abort was an operation that didn't
presented to the users who was the author of that operation so I added a
new column in Taks to store who abborted it and I'm presenting in the
view.