Fix namespaced lerna packages
## What
This PR updates @goodforonefare's lerna support so that the `assert` for lerna independent packages no longer breaks when you're using namespaced npm packages.

## How / Why
Previously we assumed tags would look like `mypackage@1.1.1`, and split on `@` to separate the version from the name.  This presents a problem as namespaced imports look like `@shopify/mypackage@1.1.1`.

To get around this we simply fork on encountering an `@` as the first character.

## Caveats
This PR still assumes some weird stuff. 
- assumes packages are all in a `packages` directory
- assumes packages are not nested in sub-folders inside packages
- assumes no slashes in the name of a package other than the namespace slash

If we wanted to make this really robust we'd probably need to parse out the right directory from `lerna.json`, and do some kind of regex approach instead of what we're doing now to parse the tags. At that point I'd probably want to rewrite this bash script in ruby.

I think it's fine for now though, as deviating from those assumptions seems wrong.