Fix stream output polling after Rails 5 defaulted to the root key being included in JSON serializers
I can't find the commit exactly where this happened, but on my fresh shipipt install with active_model_serializers at v0.9.7, the root key is included in the JSON serialization for a task by default. I imagine this wasn't the case before, and so the OutputStream that is responsible for polling the task output was expecting to find it's data in the root level keys of the response instead of under a root key for the object.

This updates the task stream to work on fresh installs of shipit with Rails 5, but I don't understand why this wasn't an issue for Shopify. Does shopify have custom AMS config in their shipit deploy?