Prevent rollback of task state transition
When an exception happens during a state transition the database
transaction is rolled back. When Redis is not responding a failure occur
and the task state is modified to error, however during the transition
to error the code tries to dispatch a job, which may fail, causing the
transition to rollback.