fix CVE-2015-9284 per omniauth/omniauth#809
The request phase of the OmniAuth Ruby gem is vulnerable to Cross-Site Request Forgery when used as part of the Ruby on Rails framework, allowing accounts to be connected without user intent, user interaction, or feedback to the user. This permits a secondary account to be able to sign into the web application as the primary account.

same fix as merged in Shopify/omniauth-identity#8
@see https://nvd.nist.gov/vuln/detail/CVE-2015-9284