Improve 'Commits not in deploy' presentation
In the current state, the commit not included in the deploy are
presented inside a section with the header that says "Commits included
in this deploy", which creates the wrong impression on the user.

Moreover, these not-included commits are listed above the "Create
Deploy" button, giving the impression that they are a part of the
commits that will be included in the action.

This commit moves the "Commits not included in this deploy" section
outside and below the "Create Deploy" button so that it is visually
separate from where the action is expected to happen. The information is
still being displayed in a de-emphasized manner, and the title of the
new section is very clear, so it should not create extra user
confusion.