Drops duplicate Continuous Delivery jobs
This job does not explicitly set the lock timeout, and so uses the default of `0`. It also defaults to `on_duplicate :retry`, despite being executed every minute by the cron job. It is highly likely that many of these jobs are duplicates that are being retried, though this is entirely unnecessary because the job runs every minute to catch up anyway, causing a very high job load and many reported exceptions to exception tracking services.