Merge upstream master from Shopify/shipit-engine (#18)
* Release 0.30.0

* Update changelog with 0.30.0 version entry

* Render shipit.yml links as a menu-list in the nav

For stacks with more than one or two links the navigation bar starts
to wrap links. We _might_ prevent this by rendering the links as a
sub-list instead.

* Update omniauth-github

* Add ability to perform variable substitution to stack links

...when at least one task has been performed on the stack - IE when
we have something we can use to grab the environment variables.

This maintains the previous functionality of the `shipit.yml` to
declare static links for stacks like:

```yml
links:
  logs: https://logs.application.com
```

but also affords consumers the ability to define template strings in
the `shipit.yaml` like:

```yml
links:
  logs: https://logs.${ENVIRONMENT}.application.com
```

Which, if the stack's environment is set to `production` will render a
link to `https://logs.production.application.com` in the stack's header.

Additionally, allow host applications to add environment values to the
stack. All stack env values are also exposed to the `Stack#links`
method for variable substitution.

The host applications can provide their own values to the stack's
environment by intercepting messages dispatched to the `Stack#env` -
for example, a host application might do something like this in an
initializer:

config/initializer/shipit.yml
```ruby
ActiveSupport::Reloader.to_prepare do
...
  module EnvironmentExtensions
    def env
      super
        .merge(
         # ... environment values from host application
        )
    end
  end

  require 'shipit/stack'
  Shipit::Stack.prepend(EnvironmentExtensions)
end
```

Co-authored-by: Jean Boussier <jean.boussier@gmail.com>
Co-authored-by: Jose Luis Salas <josacar@users.noreply.github.com>
Co-authored-by: Jean byroot Boussier <jean.boussier+github@shopify.com>