[GOG-153] Prevent erroneously marking deployments as dead soon after they start (#27)
* Prevent immediately marking newly started deployments as dead

There appears to be a race condition between the time when the
PerformTaskJob marks a task as "running" and when the newly added
RemoveDeadTasks job attempts to reap dead tasks.

In some cases a deployment is almost immediately marked as dead after
it starts. We believe this is because there critical section in which
PerformTaskJob marks a task as running and ReapDeadTaskJob is run in
which newly a task can be transitioned to a "running" status, but not
yet marked alive in which a newly created task will be considered a
"dead task" by the ReapDeaTasks job.

This changeset aims a two pronged approach at preventing this
race. First it makes the ReapDeadTasksJob less aggressive in its
consideration of tasks as "dead" - it will now only consider tasks in
a "running" or "aborted" state which are missing an "alive" monitor
AND are 5 at least minutes old.

Second, when PerformTaskJob starts working a task the task will ping
the "alive" monitor as soon as the task is marked as running. The
implementation of reporting that tasks are "alive" is somewhat
imperfect, in that it it doesn't happen until we attempt to capture
command output. There are therefore several things which happen
between starting the task and the first command happening, namely one
or two checks that the git repo is already checked out and a wait on a
cache lock, whose timeout is also 15 seconds. This means a large
opportunity for the "alive" status to expire. We should probably
ensure that the task is considered alive for a period long enough to
span this (maybe 30 seconds rather than 15), as well as ping after
each check of `fetched?`.

References
----------

- [Task marked as running, but not yet
alive](https://github.com/Shopify/shipit-engine/blob/d48ceb4ffc21afbdf2701ce8c3a8936268afc3d5/app/jobs/shipit/perform_task_job.rb#L18)
- [Task first marked as alive](https://github.com/Shopify/shipit-engine/blob/d48ceb4ffc21afbdf2701ce8c3a8936268afc3d5/app/jobs/shipit/perform_task_job.rb#L81)