Add ReviewStack into shipit engine (#37)
* Add ReviewStack into shipit engine

Makes ReviewStacks a 1st class citizen of shipit-engine. Additionally
serves as a place to start pulling the review stack functionality from
our host application down into shipit-engine.

We've not pulled functionality from the host application into this
body of work to minimize the risk to our existing users - IE let's
tinker with the engine while the car's running down the highway, but
let's leave the wheels alone for now. That is to say, there is much
more work coming.

* Fix review stack settings pages

When accessing the review stack's setting page the ActiveRecord
"record identification" system takes over and attempts to route form
submits to `review_stack_path`. At this time there are no specific
forms for `review_stacks` so we don't need this level of helpfulness
from rails. Rather than imply attributes about the forms from the
model, for now, lets be explicit about what we want.