Rename Shipit::PullRequest to Shipit::MergeRequest
Implements the proposal of Shopify/shipit-engine issue #1093 by
renaming the concept of the `Shipit::PullRequest` to
`Shipit::MergeRequest`. Consideration was given to maintaining the
name "PullRequest" where shipit-engine is interacting directly with
the Github API - for example, in the `Shipit.github.api client`. In
this way the `Shipit::MergeRequest` is related to Github Pull
Requests, but not necessarily _of_ a Github Pull Request.

The goal is to make way for a more direct modeling of a Github Pull
Request in shipit-engine - `Shipit::PullRequest` - which like
`Shipit::Commit` models the Github concept. The plan is to use this
closer modeling of the Github Pull Request to finish preparations for
submission of our "Review Stack" feature to `Shopify/shipit-engine`.

Notes
-------

This does include non-backward compatible HTTP API changes to
shipit-engine - the `pull_requests` collection resources have been
renamed to `merge_requests`.

References
----------

- https://github.com/Shopify/shipit-engine/issues/1093