Add head commit to pull requests (#56)
We use this in our host application to determine if the currenty
deployment would put the Review Stack in a "current with the head
of the pull request" state - if so, we send notifications to the user
who opened the pull request indicating that their instance is
"ready". This was missed in the original PullRequest implementation.