Merge shopify master into pr review stacks branch (#65)
* Use api responses directly instead of through last_response

* Move exception handling

* Don't set response if no data

* Add mini_racer in development to not depend on node or other system wide JS engine

* Preload Stack#repository by default to fix several N+1

* Fix mocha deprecations

* Add a power user endpoint to list all stacks of a repo

* Implement a stacks#update API

* Fix api/stacks#update route

* Save the stack before locking if only the spec has changed

* Local dev fixes

* Add dedicated view for tasks

* Add rollback api

* Add frozen string literal comment

* Make the latest deployed reference branch opt-in

Fix: #955

* Fix timeline page

* Abort active deploy on rollback with force

* Perform rollback after abort

Use mechanism for triggering revert after abort to trigger a rollback to
a specific deploy.

* Update test to use until commit id rather than deploy

* Update npm-lerna.md

* Adapt merge status to new GitHub UI

Github changed the width of their page which is why the "Add to merge
queue" button looks out of place now. I removed the max-width constraint
as this gave me the right results when playing around in the chrome web
inspector.

* Remove capitilisation from ShipIt repo listing

* Update dependencies and test Ruby 2.7

* Workaround the active_model_serializer lookup bug

Long story short the constant lookip semantic changed slightly in
2.7.0+ and 2.6.6+.

Previously `Shipit::Shipit::Commit` would evaluate to `::Shipit:Commit`.

Because of this, badly configured AMS would kinda work. Now it need
to be more correct.

* Update bundler on CI

* Release 0.32.0

* rubocop formatting

* changed stack query

* Fix query and add test for an edge-case.

* Disable the Style/PerlBackrefs cop

* Avoid loading useless parts of Rails in development

* Rename Shipit::PullRequest to Shipit::MergeRequest

Implements the proposal of Shopify/shipit-engine issue #1093 by
renaming the concept of the `Shipit::PullRequest` to
`Shipit::MergeRequest`. Consideration was given to maintaining the
name "PullRequest" where shipit-engine is interacting directly with
the Github API - for example, in the `Shipit.github.api client`. In
this way the `Shipit::MergeRequest` is related to Github Pull
Requests, but not necessarily _of_ a Github Pull Request.

The goal is to make way for a more direct modeling of a Github Pull
Request in shipit-engine - `Shipit::PullRequest` - which like
`Shipit::Commit` models the Github concept. The plan is to use this
closer modeling of the Github Pull Request to finish preparations for
submission of our "Review Stack" feature to `Shopify/shipit-engine`.

Notes
-------

This does include non-backward compatible HTTP API changes to
shipit-engine - the `pull_requests` collection resources have been
renamed to `merge_requests`.

References
----------

- https://github.com/Shopify/shipit-engine/issues/1093

* Don't redirect on enabling emergency mode

* Removes capitalization of environments

A-la https://github.com/Shopify/shipit-engine/pull/1091

* Serialize rollback once aborted to value

* Revert the pull_request -> merge_request renaming in CommitSerializer [ref #1095]

* Merge shopify master (#64)

* Don't redirect on enabling emergency mode

* Removes capitalization of environments

A-la https://github.com/Shopify/shipit-engine/pull/1091

* Serialize rollback once aborted to value

* Revert the pull_request -> merge_request renaming in CommitSerializer [ref #1095]

Co-authored-by: Jack Li <jack.li@shopify.com>
Co-authored-by: Ben Langfeld <blangfeld@powerhrg.com>
Co-authored-by: Jean byroot Boussier <jean.boussier+github@shopify.com>
Co-authored-by: Jean Boussier <jean.boussier@gmail.com>

Co-authored-by: Jack Li <jack.li@shopify.com>
Co-authored-by: Jean Boussier <jean.boussier@gmail.com>
Co-authored-by: Darren Worrall <darren.worrall@shopify.com>
Co-authored-by: Darren Worrall <DazWorrall@users.noreply.github.com>
Co-authored-by: Kaelig Deloumeau-Prigent <kaelig@users.noreply.github.com>
Co-authored-by: Robin Brandt <robin.brandt@shopify.com>
Co-authored-by: Robin Brandt <robinbrandt@users.noreply.github.com>
Co-authored-by: Alex Page <alex@alexpage.com.au>
Co-authored-by: Jean byroot Boussier <jean.boussier+github@shopify.com>
Co-authored-by: Ruidan Ji <ruidan.ji@shopify.com>
Co-authored-by: Jonathan Geiger <jonathan.geiger@shopify.com>
Co-authored-by: rdji123 <36777675+rdji123@users.noreply.github.com>
Co-authored-by: Ben Langfeld <blangfeld@powerhrg.com>