Remove ReviewStacks awaiting provision from queue when archiving (#1106)
* Remove ReviewStacks awaiting provision from queue when archiving

Why
===

When a `ReviewStack` is created and queued for provisioning, it's
possible that the Pull Request is closed or its labels are adjusted in
such a way that the `ReviewStack` is considered "Archived" before the
`ReviewStackProvisioningQueue#work` method has a chance to process
queue. In these rare cases, we want to remove the `ReviewStack` from
the provisioning queue. While leaving the `ReviewStack` in the
`ReviewStackProvisioningQueue` doesn't cause major misbehavior of the
system, it does leave the data in a somewhat confusing state and can
unnecessarily grow the amount of `ReviewStack`s the
`ReviewStackProvisioningQueue` must process on each invocation of
`#work`.

* Move provisioning en/dequeue checks to ReviewStack

Responding to code review feedback:

- https://github.com/Shopify/shipit-engine/pull/1106/files#r499930355