Use the monotonic clock for task timing
Using Time.now MAY be unreliable. On some systems, Time.now uses the
`gettimeofday` system call. The `gettimeofday` documentation states:

> The time returned by gettimeofday() is affected by discontinuous jumps
in the system time (e.g., if the system administrator manually changes
the system time).

IE this returns "wall time" of the system clock which MAY change
unexpectedly - for example if the system uses NTP to set time and the
clock adjusts during the execution of a task. Time.now is prone to
jumping around an inconsistent timeline.

Instead, this uses the system's monotonic clock to ensure that "time" is
consistently measured between the "start" and "stop" of a task.

References
----------

- http://man7.org/linux/man-pages/man2/gettimeofday.2.html
- https://blog.dnsimple.com/2018/03/elapsed-time-with-ruby-the-right-way/
- https://ruby-doc.org/core-2.5.3/Process.html#method-c-clock_gettime