Remove vestigial task_execution_strategy classes
These are remnant of the original implementation of the "Swap-able Task Execution
Strategy" work. The upstream contribution saw the removal of the
execution strategy registry however the registry was already included in
our working branch. This cleans up the now unused bits still hanging about
in our fork so we're more closely aligned with the upstream branch.

References
----------

- https://github.com/Shopify/shipit-engine/pull/1107