Merge shopify master into working (#85)
* Update mimemagic

* Disable GitHub actions fast failing

* Fix Ruby 3.0 keyword argument passing in DeliverHookJob

* Fix the MySQL test suite

* Minor rails update

* Support for sending signed webhooks with the secret key (#1150)

* Use Hook's secret in webhook delivery

The hook model has a "secret" column but the value is currently unused.
This change will send that string as the HTTP header X-Shipit-Secret in
the webhook, which can be coordinated with the recipient to provide
authentication.

* Introduce test for delivery

* Provide a secret to be sent for the test

When secret is unset, our HTTP library will default to not sending the
header since it's value is nil. But the test library converts that nil
to an empty string and expects it to be present.

* Switch to a signature-based approach using sha256

* Don't send signature header if secret is empty

* Use blank to handle both '' and nil

* Use inferred default stack branch for stacks

* Fix bundler deprecation

* Support multiple GH organisations (#1151)

* Add multi Github application support

* Update tests

Co-authored-by: Marvin Frick <marvin.frick@shopify.com>

* Update documentation

* Implement feedback

* Delegate github_api to stack

* Clean up tests requiring stack for GithubApp

Co-authored-by: Riley Shott <riley.shott@shopify.com>

* Regenerate the Gemfile.lock

* Fix rubocop

* Improve the loading of GitHub apps from secrets

* Force standardization of github config keys

* Move symbolization of Rails.application.secrets up the stack

* Fix anonymous user handling in tasks and commits

* Fix WebhookController calls

* Add logging to WebhookController#verify_signature

* Add indifferent access support for searching through Github applications

* Handle more payload types in WebhookController calls

* GitHub push events only update active stacks

We are observing multiple archived stacks recreate their local file
caches. When this occurs in repositories which have
high-review-stack-volume and many files we observe large amounts of disk
utilization - especially inode space utilization. This has deleterious
effects, causing CacheDeploySpec jobs, Deployments, and anything else
which has an opportunity to interact with the local file cache to
encounter filesystem errors like:

```
$ git clone --quiet --local --origin cache [snip]
fatal: cannot create directory at '[snip]': No space left on device
warning: Clone succeeded, but checkout failed.
You can inspect what was checked out with 'git status'
and retry the checkout with 'git checkout -f HEAD'
```

There appear to be more than one cause for these old, archived stack
recreating their local file caches, but the one with which this PR
concerns itself is as follows. When `PushHandler#process` receives a
"Commit" GitHub webhook payload whenever a GitHub branch receives a push
from the developer.

It finds every stack whose `branch` attribute matches that of the of the
GitHub webhook's Commit payload. Then Shipit calls `sync_github` on the
each of these stacks[^1]. `sync_github` pulls the missing commit information
for each stack and creates Shipit::Commits for them. Then the
`GithubSyncJob` enqueues a `CacheDeploySpecJob` for the stack[^2]. The
`CacheDeploySpecJob` then calls
`Shipit::StackCommands#with_temporary_directory` to sync the local git
cache and rebuild stack's DeploySpec.

This is where we end up rebuilding the on-disk cache for archived
stacks. Keep in mind that in this process, we never concern ourselves
with the `archived?` status of the stack. If an archived stack shares a
`#branch` attribute with an active stack its on-disk file cache is
recreated here because `with_temporary_directory` will run the `fetch`
command on the stack[^3].

[^1]: https://github.com/Shopify/shipit-engine/blob/e7328e83d8e78ca878f1f5ad2cdebe2ec733854d/app/models/shipit/webhooks/handlers/push_handler.rb#L10
[^2]: https://github.com/Shopify/shipit-engine/blob/a0a826555691f0c83cf74667e6df09911ff339e4/app/jobs/shipit/github_sync_job.rb#L28
[^3]: https://github.com/Shopify/shipit-engine/blob/e7328e83d8e78ca878f1f5ad2cdebe2ec733854d/lib/shipit/stack_commands.rb#L56

* Improve sync_github_if_necessary guard readability

Co-authored-by: Jean byroot Boussier <jean.boussier+github@shopify.com>

* Improve readability of schedule_merges_if_necessary

Feedback provided in PR #1161 improved the readability of a similar
life-cycle callback - `sync_github_if_necessary`. This carries the idea
forward to - hopefully - save readers of the
`schedule_merges_if_necessary` method some mental computation.

References
----------

- https://github.com/Shopify/shipit-engine/pull/1161

Co-authored-by: Jean Boussier <jean.boussier@gmail.com>
Co-authored-by: Ryan Wilson-Perkin <ryan.wilsonperkin@shopify.com>
Co-authored-by: Gannon McGibbon <gannon.mcgibbon@shopify.com>
Co-authored-by: Jean byroot Boussier <jean.boussier+github@shopify.com>
Co-authored-by: Gannon McGibbon <gannon.mcgibbon@gmail.com>
Co-authored-by: Marv <MrMarvin@users.noreply.github.com>
Co-authored-by: Riley Shott <riley.shott@shopify.com>
Co-authored-by: Marvin Frick <marvin.frick@shopify.com>
Co-authored-by: Riley Shott <Ginja@users.noreply.github.com>