Fix "view on github" links for Review Stacks
For a `ReviewStack`, "View on GitHub" link of the `stacks/_header`
template appropriately will send users to the Pull Request's so long as
the user is NOT on a "Task" oriented view - IE the History, or Task's
show view, etc. When looking at a Review Stack's tasks the "View on
GitHub" link - less-than-optimally - sends the user to the Stack's
GitHub repository.

This seems to happen because when the `GithubUrlHelper` receives the
`ReviewStack` to determine which type of link it needs to render. The
Tasks Controller casts the `ReviewStack` to a Stack instance before this
call to the helper - meaning that the helper's test of "Does this stack
have a pull_request?" fails [^1]. This causes the helper to render a
link to the GitHub Repository.

The TasksController casts the ReviewStack so that the post-create
redirects send the user to the Task's show view - since
review_stacks_task_path is not a valid path helper. This change isolates
the cast to the post-create redirect.

[^1]: https://github.com/Shopify/shipit-engine/blob/master/app/helpers/shipit/github_url_helper.rb#L55