Fix "view on github" links for Review Stacks
For a `ReviewStack`, "View on GitHub" link of the `stacks/_header`
template appropriately will send users to the Pull Request's so long as
the user is NOT on a "Task" oriented view - IE the History, or Task's
show view, etc. When looking at a Review Stack's tasks the "View on
GitHub" link - less-than-optimally - sends the user to the Stack's
GitHub repository.

This seems to happen because when the `GithubUrlHelper` receives the
`ReviewStack` to determine which type of link it needs to render. The
Tasks Controller up-casts the `ReviewStack` to a `Stack` instance - via
the `#becomes` method - before this call to the helper - meaning that
the helper's test of "Does this stack have a pull_request?" fails [^1].
This causes the helper to render a link to the GitHub Repository.

These up-casts serve to instruct Rails' path helpers that `ReviewStack`s
and `Stacks` are the same as far as the application's routes and view
paths are concerned. This change consolidates this fact into the
`ReviewStack` itself rather than scattering this knowledge throughout
the shipit-engine controllers and views.

[^1]: https://github.com/Shopify/shipit-engine/blob/master/app/helpers/shipit/github_url_helper.rb#L55