Optionally record Review Stack latest deployed ref
If enabled, `update_latest_deployed_ref` creates a branch for each
"Stack" that shipit-engine deploys. In a Repository with high Review
Stack turnover this quickly adds up to a lot of "noise" branches in the
git object database- our flagship monorepo currently has over 3k such
branches. However, in some scenarios it is still useful to record the
"latest deployed ref" of "plain old, manually managed" stacks - in our
case we DO still want to track the latest deployed ref of our staging
and production "environments" of this flagship monorepo.

This disables recording the latest deployd ref branch for ReviewStacks